image  preview and delete

+++ app/assets/javascripts/modal.js
@@ -11,7 +11,7 @@ function modalOpen(e){
     var width = modal_item.clientWidth;
     var height = width * (9/16)
     console.log(width)
-    modal_item.innerHTML = '<iframe width="'+width+'" height="'+height+'" src="https://www.youtube.com/embed/'+movID+'?rel=0&showinfo=0" frameborder="0" allowfullscreen></iframe>';
+    modal_item.innerHTML = `<iframe width="${width}" height="${height}" src="https://www.youtube.com/embed/'+movID+'?rel=0&showinfo=0" frameborder="0" allowfullscreen></iframe>`
   }
   if (modal_open[e].dataset.live == "true") {
     const liveID = modal_open[e].dataset.id;
@@ -19,15 +19,13 @@ function modalOpen(e){
     const description = modal_open[e].dataset.description;
     modal_item.innerHTML = `<p class="modal_text">データを削除します。よろしいですか？</p>
                             <p class="modal_text note">※削除したデータの復元はできません。</p>
-                            <p class="modal_title">`+title+`</p>
-                            <p class="modal_description">`+description+`</p>
+                            <p class="modal_title">${title}</p>
+                            <p class="modal_description">${description}</p>
                             <div class="modal_selection">
                               <div class="btn cancel">キャンセル</div>
-                              <div class="btn confirm">
-                                <a rel="nofollow" data-method="delete" href="/lives/`+liveID+`">
-                                  <p>削除する</p>
-                                </a>
-                              </div>
+                              <a class="btn confirm" rel="nofollow" data-method="delete" href="/lives/${liveID}">
+                                <p>削除する</p>
+                              </a>
                             </div>`
     const cancel = document.getElementsByClassName('cancel')[0];
     cancel.addEventListener('click', function(){
